+++ starsky-tools/build-tools/lib/https-get.js
@@ -1,14 +1,17 @@
 const https = require("https");
 
-
 async function httpsGet(url, authorizationHeader = "") {
 	return new Promise((resolve, reject) => {
-		httpsGetInternal(url, authorizationHeader).then(resolve).catch(()=> {
-			httpsGetInternal(url, authorizationHeader).then(resolve).catch((err)=> {
-				console.log('failed');
-				reject(err)
-			})
-		})
+		httpsGetInternal(url, authorizationHeader)
+			.then(resolve)
+			.catch(() => {
+				httpsGetInternal(url, authorizationHeader)
+					.then(resolve)
+					.catch((err) => {
+						console.log("failed");
+						reject(err);
+					});
+			});
 	});
 }
 
@@ -48,11 +51,13 @@ async function httpsGetInternal(url, authorizationHeader = "") {
 			// response finished, resolve the promise with data
 			res.on("end", () => {
 				try {
-					const parsedResponse = JSON.parse(response)
+					const parsedResponse = JSON.parse(response);
+					parsedResponse.headers = res.headers;
 					resolve(parsedResponse);
 				} catch (error) {
-					error.url = requestOptions.host + requestOptions.path
-					error.statusCode = res.statusCode
+					error.url = requestOptions.host + requestOptions.path;
+					error.statusCode = res.statusCode;
+					error.headers = res.headers;
 					reject(error);
 				}
 			});