+++ starsky/starsky.foundation.database/Notifications/NotificationQuery.cs
@@ -75,7 +75,7 @@ namespace starsky.foundation.database.Notifications
 		public Task<NotificationItem> AddNotification<T>(ApiNotificationResponseModel<T> content)
 		{
 			var stringMessage = JsonSerializer.Serialize(content,
-				DefaultJsonSerializer.CamelCase);		
+				DefaultJsonSerializer.CamelCaseNoEnters);		
 			return AddNotification(stringMessage);
 		}