+++ starsky/starsky.foundation.storage/Storage/StorageHostFullPathFilesystem.cs
@@ -41,7 +41,7 @@ namespace starsky.foundation.storage.Storage
 			}
 
 			var lastWrite = type == FolderOrFileModel.FolderOrFileTypeList.File ? 
-				File.GetLastWriteTime(path).ToUniversalTime() : Directory.GetLastWriteTime(path).ToUniversalTime();
+				File.GetLastWriteTime(path) : Directory.GetLastWriteTime(path);
 			
 			var size = type == FolderOrFileModel.FolderOrFileTypeList.File ? 
 				new FileInfo(path).Length : -1;
@@ -216,6 +216,7 @@ namespace starsky.foundation.storage.Storage
 				if ( maxRead < 1 ) return fileStream;
 				
 				byte[] buffer = new byte[maxRead];
+				// ReSharper disable once MustUseReturnValue
 				fileStream.Read(buffer, 0, maxRead);
 				fileStream.Close();
 				return new MemoryStream(buffer);
@@ -445,6 +446,25 @@ namespace starsky.foundation.storage.Storage
 	            RecurseFind(s, list);
             }
         }
+
+		public DateTime SetLastWriteTime(string path, DateTime? dateTime = null)
+		{
+			if ( dateTime?.Year == null || dateTime.Value.Year <= 2000 )
+			{
+				dateTime =  DateTime.Now;
+			}
+			var type = IsFolderOrFile(path);
+
+			if ( type == FolderOrFileModel.FolderOrFileTypeList.File )
+			{
+				File.SetLastWriteTime(path, dateTime.Value);
+			}
+			else
+			{
+				Directory.SetLastWriteTime(path, dateTime.Value);
+			}
+			return dateTime.Value;
+		}
 	}
 
 }