+++ starsky/starsky.foundation.storage/Storage/StorageThumbnailFilesystem.cs
@@ -173,5 +173,10 @@ namespace starsky.foundation.storage.Storage
 		{
 			throw new System.NotImplementedException();
 		}
+
+		public DateTime SetLastWriteTime(string path, DateTime? dateTime = null)
+		{
+			return dateTime ?? DateTime.Now;
+		}
 	}
 }