+++ starsky/starsky.foundation.writemeta/JsonService/FileIndexItemJsonParser.cs
@@ -2,6 +2,8 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using starsky.foundation.database.Models;
 using starsky.foundation.platform.Helpers;
+using starsky.foundation.platform.JsonConverter;
+using starsky.foundation.platform.Models;
 using starsky.foundation.storage.Helpers;
 using starsky.foundation.storage.Interfaces;
 
@@ -17,17 +19,21 @@ namespace starsky.foundation.writemeta.JsonService
 		}
 		
 		/// <summary>
-		/// Write FileIndexItem to IStorage
+		/// Write FileIndexItem to IStorage .meta.json file
 		/// </summary>
 		/// <param name="fileIndexItem">data object</param>
-		/// <returns>void</returns>
+		/// <returns>Completed Task</returns>
 		public async Task WriteAsync(FileIndexItem fileIndexItem)
 		{
-			var jsonOutput = JsonSerializer.Serialize(fileIndexItem, new JsonSerializerOptions
+			var jsonOutput = JsonSerializer.Serialize(new MetadataContainer
 			{
-				WriteIndented = true, 
-			});
-			var jsonSubPath = JsonSidecarLocation.JsonLocation(fileIndexItem.ParentDirectory, fileIndexItem.FileName);
+				Item = fileIndexItem
+			}, DefaultJsonSerializer.CamelCase);
+			
+			var jsonSubPath = JsonSidecarLocation.JsonLocation(
+				fileIndexItem.ParentDirectory, 
+				fileIndexItem.FileName);
+			
 			await _iStorage.WriteStreamAsync(
 				PlainTextFileHelper.StringToStream(jsonOutput), jsonSubPath);
 		}
@@ -43,9 +49,13 @@ namespace starsky.foundation.writemeta.JsonService
 			// when sidecar file does not exist
 			if ( !_iStorage.ExistFile(jsonSubPath) ) return fileIndexItem;
 			
-			var returnFileIndexItem = new DeserializeJson(_iStorage).Read<FileIndexItem>(jsonSubPath);
-			returnFileIndexItem.Status = FileIndexItem.ExifStatus.ExifWriteNotSupported;
-			return returnFileIndexItem;
+			var returnContainer = new DeserializeJson(_iStorage).Read<MetadataContainer>(jsonSubPath);
+			
+			// in case of invalid json
+			returnContainer.Item ??= fileIndexItem;
+			
+			returnContainer.Item.Status = FileIndexItem.ExifStatus.ExifWriteNotSupported;
+			return returnContainer.Item;
 		}
 	}