+++ starsky/starskytest/starsky.foundation.storage/Helpers/DeserializeJsonTest.cs
@@ -1,5 +1,7 @@
+using System;
 using System.IO;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using starsky.foundation.database.Models;
 using starsky.foundation.storage.Helpers;
 using starskytest.FakeMocks;
 
@@ -26,7 +28,36 @@ namespace starskytest.starsky.foundation.storage.Helpers
 			Assert.AreEqual("Title", result.Title);
 			Assert.AreEqual(true, result.ShowButtons);
 		}
-		
+
+		[TestMethod]
+		public void Item()
+		{
+			const string input = "{  \"$id\": \"https://docs.qdraw.nl/schema/meta-data-container.json\",  " +
+			                     "\"$schema\": \"https://json-schema.org/draft/2020-12/schema\",  \"item\": {    " +
+			                     "\"filePath\": \"/test.jpg\",    \"fileName\": \"test.jpg\",    \"fileHash\": \"Test\",   " +
+			                     " \"fileCollectionName\": \"test\",    \"parentDirectory\": \"/\",    \"isDirectory\": false,   " +
+			                     " \"tags\": \"test\",    \"status\": \"ExifWriteNotSupported\",    \"description\": \"Description\",    " +
+			                     "\"title\": \"Title\",    \"dateTime\": \"2020-01-01T00:00:00\",    \"addToDatabase\": \"2020-01-01T00:00:00\",    " +
+			                     "\"lastEdited\": \"2020-01-01T00:00:00\",    \"latitude\": 50,    \"longitude\": 5,    \"locationAltitude\": 1,   " +
+			                     " \"locationCity\": \"LocationCity\",    " +
+			                     "\"locationState\": \"LocationState\",    \"locationCountry\": \"LocationCountry\",    " +
+			                     "\"locationCountryCode\": null,    \"colorClass\": 2,    \"orientation\": \"Rotate180\",    " +
+			                     "\"imageWidth\": 100,    \"imageHeight\": 140,    \"imageFormat\": \"jpg\",    " +
+			                     "\"collectionPaths\": [],    \"sidecarExtensionsList\": [],    \"aperture\": 1,    " +
+			                     "\"shutterSpeed\": \"10\",    \"isoSpeed\": 1200,    \"software\": \"starsky\",    " +
+			                     "\"makeModel\": \"test|test|\",    \"make\": \"test\",    \"model\": \"test\",    " +
+			                     "\"lensModel\": \"\",    \"focalLength\": 200,    \"size\": 0,    " +
+			                     "\"imageStabilisation\": \"Unknown\",    \"lastChanged\": []  }}";
+			var fakeStorage = new FakeIStorage();
+			fakeStorage.WriteStream(
+				PlainTextFileHelper.StringToStream(input), "/test.json");
+			
+			var itemJsonParser = new DeserializeJson(fakeStorage);
+			var result = itemJsonParser.Read<MetadataContainer>("/test.json");
+
+			Assert.AreEqual("/test.jpg", result.Item.FilePath);
+		}
+
 		[TestMethod]
 		[ExpectedException(typeof(FileNotFoundException))]
 		public void ReadTest_NotFound()