+++ starsky/starskytest/starsky.foundation.sync/SyncServices/SyncSingleFileTest.cs
@@ -268,7 +268,7 @@ namespace starskytest.starsky.foundation.sync.SyncServices
 		{
 			const string currentFilePath = "/test_date2.jpg";
 			_iStorageFake.FileCopy("/test.jpg", currentFilePath);
-			(_iStorageFake as FakeIStorage)!.SetDateTime(currentFilePath,DateTime.UtcNow);
+			(_iStorageFake as FakeIStorage)!.SetLastWriteTime(currentFilePath,DateTime.UtcNow);
 			
 			var (fileHash, _) = await new FileHash(_iStorageFake).GetHashCodeAsync(currentFilePath);
 				
@@ -471,7 +471,7 @@ namespace starskytest.starsky.foundation.sync.SyncServices
 		{
 			const string filePath = "/FileAlreadyExist_With_Different_LastEditedTime.jpg";
 			_iStorageFake.FileCopy("/test.jpg", filePath);
-			( _iStorageFake as FakeIStorage )?.SetDateTime(filePath,
+			( _iStorageFake as FakeIStorage )?.SetLastWriteTime(filePath,
 				_lastEditedDateTime);
 
 			var (fileHash, _) = await new FileHash(_iStorageFake).GetHashCodeAsync(filePath);